pyproject.toml
[project]
name = "procnumnodocexec"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "SQLAlchemy>=2.0",
    "aiofile>=3.9.0",
    "pyodbc>=5.1",
    "aioodbc>=0.3.0",
    "python-decouple>=3.8",
    "smbprotocol>=1.15.0",
    "tqdm>=4.0.0",
]
